hanlhn/pyproject.toml
[tool.poetry]
name = "hanlhn"
version = "0.1.0"
description = ""
authors = ["hanahh080601 <hanlehngoc080601@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
pymongo = {extras = ["srv"], version = "^4.1.1"}
pydantic = "^1.9.1"
Pillow = "^9.1.1"
timm = "^0.5.4"
pandas = "^1.4.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
